requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# 数据库和ORM
sqlalchemy==2.0.23
alembic==1.12.1

# AI和OCR
openai==1.3.7  # 用于调用阿里百炼API（兼容OpenAI格式）
requests==2.31.0
httpx==0.25.2

# 工具库
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0